refactor(store): extract port-extraction helpers to twinPortExtraction
Move extractPortsFromKubeYaml + extractPortsFromQuadletContainer +
safeParsePort out of TwinStore into a sibling pure-function module.
extractStaticPortsForService now reads as: cache check -> pushPort
closure -> kube branch -> container branch -> cache set.

Lint on twin.ts:
- file-lines warning cleared (1083 -> 974 raw lines)
- extractStaticPortsForService line warning cleared (107 -> ~50)
- extractStaticPortsForService complexity still over (was 35, now 20;
  pushPort closure is the remaining contributor)

Total warnings: 419 -> 418.
